[IMP] purchase_work_acceptance_kmitl: consolidate ribbon and purchase_order modules
Merge purchase_work_acceptance_ribbon and purchase_work_acceptance_purchase_order
into purchase_work_acceptance_kmitl to reduce module fragmentation and simplify
dependency management. All features are preserved including the ribbon widget,
purchase order navigation button, and Thai translations.
