+++ kmitl_depends_dev/__manifest__.py
@@ -24,8 +24,8 @@
                 'purchase_request_attachment', 'purchase_request_hide_create_po_button', 'purchase_request_operating_unit_access_all',
                 'purchase_sequence_kmitl', 'purchase_request_sarabun', 'purchase_request_sequence_kmitl',
                 'purchase_request_ux_kmitl', 'purchase_work_acceptance_invoice_plan_deliverables', 'purchase_work_acceptance_invoice_plan_usability',
-                'purchase_work_acceptance_kmitl', 'purchase_work_acceptance_portal', 'purchase_work_acceptance_purchase_order',
-                'purchase_work_acceptance_ribbon', 'stock_inventory_department', 'stock_inventory_kmitl', 'stock_inventory_restriction',
+                'purchase_work_acceptance_kmitl', 'purchase_work_acceptance_portal',
+                'stock_inventory_department', 'stock_inventory_kmitl', 'stock_inventory_restriction',
                 'stock_scrap_reason_text', 'stock_scrap_responsible_user', 'thai_date_utils', 'web_kmitl', 'web_widget_helper_text', 'web_theme_classic_extended', 'kmitl_project',
                 'hr_department_short_name_tracking', 'hr_department_tracking', 'procurement_type_no_security', 'purchase_invoice_plan_kmitl',
                 'purchase_guarantee_account_payment', 'purchase_guarantee_attachment', 'purchase_guarantee_bid_guarantee_kmitl',
