--- purchase_work_acceptance_purchase_order/i18n/th.po
@@ -1,35 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* purchase_work_acceptance_purchase_order
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 16.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2025-09-22 08:18+0000\n"
-"PO-Revision-Date: 2025-09-22 15:18+0700\n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"Language: th\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Plural-Forms: nplurals=1; plural=0;\n"
-"X-Generator: Poedit 3.6\n"
-
-#. module: purchase_work_acceptance_purchase_order
-#: model_terms:ir.ui.view,arch_db:purchase_work_acceptance_purchase_order.view_work_acceptance_form
-msgid "<span class=\"o_stat_text\">Purchase Order</span>"
-msgstr "สัญญา/ใบสั่งซื้อ/จ้าง"
-
-#. module: purchase_work_acceptance_purchase_order
-#. odoo-python
-#: code:addons/purchase_work_acceptance_purchase_order/models/work_acceptance.py:0
-#, python-format
-msgid "Purchase Order"
-msgstr ""
-
-#. module: purchase_work_acceptance_purchase_order
-#: model:ir.model,name:purchase_work_acceptance_purchase_order.model_work_acceptance
-msgid "Work Acceptance"
-msgstr ""
