--- purchase_work_acceptance_purchase_order/models/__init__.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-from . import work_acceptance
