--- purchase_work_acceptance_purchase_order/views/work_acceptance_views.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-  <odoo>
-      
-    <!-- View work.acceptance form -->
-    <record id="view_work_acceptance_form" model="ir.ui.view">
-        <field name="name">view.work.acceptance.form</field>
-        <field name="model">work.acceptance</field>
-        <field name="inherit_id" ref="purchase_work_acceptance_late_fines.view_work_acceptance_fines_late_tab_form"/>
-        <field name="arch" type="xml">
-            <xpath expr="//div[@name='button_box']" position="inside">
-              <button name="action_view_purchase_order"
-                        type="object"
-                        class="oe_stat_button"
-                        icon="fa-file-text-o"
-                        attrs="{'invisible': [('purchase_id', '=', False)]}">
-                    <div class="o_field_widget o_stat_info">
-                        <span class="o_stat_text">Purchase Order</span>
-                    </div>
-              </button>
-            </xpath>
-        </field>
-    </record>
-
-  </odoo>
