--- purchase_work_acceptance_ribbon/__manifest__.py
@@ -1,16 +1,0 @@
-# -*- coding: utf-8 -*-
-{
-    'name': 'Purchase Work Acceptance Ribbon',
-    'version': '16.0.1.0.0',
-    'summary': """ Purchase Work Acceptance Ribbon Summary """,
-    "category": "KMITL",
-    "author": "Aginix Technologies",
-    "website": "https://github.com/aginix/kmitl",
-    'depends': ['purchase_work_acceptance'],
-    "data": [
-        "views/work_acceptance_views.xml"
-    ],
-    'installable': True,
-    'auto_install': False,
-    'license': 'LGPL-3',
-}
