--- purchase_work_acceptance_ribbon/views/work_acceptance_views.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-  <odoo>
-    <record id="view_work_acceptance_form" model="ir.ui.view">
-        <field name="name">view.work.acceptance.form (Ribbon)</field>
-        <field name="model">work.acceptance</field>
-        <field name="inherit_id" ref="purchase_work_acceptance.view_work_acceptance_form"/>
-        <field name="arch" type="xml">
-            <xpath expr="//div[hasclass('oe_title')]" position="before">
-                <widget
-                    name="web_ribbon"
-                    title="Accepted"
-                    bg_color="bg-success"
-                    attrs="{'invisible': [('state', '!=', 'accept')]}"
-                />
-            </xpath>
-        </field>
-    </record>
-  </odoo>
